+++ test/accessor/dataframe/test_fillna_regression.py
@@ -7,7 +7,7 @@ from dtoolkit.accessor.dataframe import fillna_regression  # noqa
 
 
 @pytest.mark.parametrize(
-    "df, method, X, y, how, kwargs, expected",
+    "df, method, columns, how, kwargs, expected",
     [
         (
             pd.DataFrame(
@@ -21,8 +21,7 @@ from dtoolkit.accessor.dataframe import fillna_regression  # noqa
                 columns=["x1", "x2", "y"],
             ),
             linear_model.LinearRegression,
-            ["x1", "x2"],
-            "y",
+            {"y": ["x1", "x2"]},
             "na",
             {},
             pd.DataFrame(
@@ -48,8 +47,7 @@ from dtoolkit.accessor.dataframe import fillna_regression  # noqa
                 columns=["x1", "x2", "y"],
             ),
             linear_model.Ridge,
-            ["x1", "x2"],
-            "y",
+            {"y": ["x1", "x2"]},
             "na",
             {},
             pd.DataFrame(
@@ -75,8 +73,7 @@ from dtoolkit.accessor.dataframe import fillna_regression  # noqa
                 columns=["x1", "x2", "y"],
             ),
             linear_model.LinearRegression,
-            ["x1", "x2"],
-            "y",
+            {"y": ["x1", "x2"]},
             "all",
             {},
             pd.DataFrame(
@@ -102,8 +99,7 @@ from dtoolkit.accessor.dataframe import fillna_regression  # noqa
                 columns=["x1", "x2", "y"],
             ),
             linear_model.LinearRegression,
-            "x1",
-            "y",
+            {"y": "x1"},
             "na",
             {},
             pd.DataFrame(
@@ -129,8 +125,7 @@ from dtoolkit.accessor.dataframe import fillna_regression  # noqa
                 columns=["x1", "x2", "y"],
             ),
             linear_model.LinearRegression,
-            "x2",
-            "y",
+            {"y": "x2"},
             "na",
             {},
             pd.DataFrame(
@@ -156,8 +151,7 @@ from dtoolkit.accessor.dataframe import fillna_regression  # noqa
                 columns=["x1", "x2", "y"],
             ),
             linear_model.LinearRegression,
-            ["x1"],
-            "y",
+            {"y": ["x1"]},
             "all",
             {},
             pd.DataFrame(
@@ -183,8 +177,7 @@ from dtoolkit.accessor.dataframe import fillna_regression  # noqa
                 columns=["x1", "x2", "y"],
             ),
             linear_model.LinearRegression,
-            pd.Index(["x1"]),
-            "y",
+            {"y": pd.Index(["x1"])},
             "all",
             {},
             pd.DataFrame(
@@ -210,8 +203,7 @@ from dtoolkit.accessor.dataframe import fillna_regression  # noqa
                 columns=["x1", "x2", "y"],
             ),
             tree.DecisionTreeRegressor,
-            ["x1", "x2"],
-            "y",
+            {"y": ["x1", "x2"]},
             "na",
             dict(criterion="friedman_mse", splitter="best"),
             pd.DataFrame(
@@ -225,10 +217,62 @@ from dtoolkit.accessor.dataframe import fillna_regression  # noqa
                 columns=["x1", "x2", "y"],
             ),
         ),
+        (
+            pd.DataFrame(
+                [
+                    [1, 1, None],
+                    [1, 2, 8],
+                    [2, 2, 9],
+                    [2, 3, 11],
+                    [3, 5, None],
+                ],
+                columns=["x1", "x2", "y"],
+            ),
+            tree.DecisionTreeRegressor,
+            {"y": "x2", "y": ["x1"], "y": ["x1", "x2"]},  # noqa: F601
+            "na",
+            {},
+            pd.DataFrame(
+                [
+                    [1, 1, 6],
+                    [1, 2, 8],
+                    [2, 2, 9],
+                    [2, 3, 11],
+                    [3, 5, 11],
+                ],
+                columns=["x1", "x2", "y"],
+            ),
+        ),
+        (
+            pd.DataFrame(
+                [
+                    [1, 1, None, 6],
+                    [1, 2, 8, 8],
+                    [2, 2, 9, 9],
+                    [2, 3, 11, 11],
+                    [3, 5, None, None],
+                ],
+                columns=["x1", "x2", "y1", "y2"],
+            ),
+            tree.DecisionTreeRegressor,
+            {"y2": ["x1"], "y1": ["x1", "x2"]},
+            "na",
+            {},
+            pd.DataFrame(
+                [
+                    [1, 1, 6, 6],
+                    [1, 2, 8, 8],
+                    [2, 2, 9, 9],
+                    [2, 3, 11, 11],
+                    [3, 5, 11, 11],
+                ],
+                columns=["x1", "x2", "y1", "y2"],
+            ),
+        ),
     ],
 )
-def test_work(df, method, X, y, how, kwargs, expected):
-    result = df.fillna_regression(method, X, y, how=how, **kwargs)
+def test_work(df, method, columns, how, kwargs, expected):
+    result = df.fillna_regression(method, columns, how=how, **kwargs)
 
     assert result.equals(result)
 
@@ -245,4 +289,4 @@ def test_error():
         columns=["x1", "x2", "y"],
     )
     with pytest.raises(ValueError):
-        df.fillna_regression(linear_model.MultiTaskElasticNet, "x1", "y", how="blah")
+        df.fillna_regression(linear_model.MultiTaskElasticNet, {"y": "x1"}, how="blah")